osd/scheduler: Factor in num of op threads to determine OSD shards' BW capacity
This change impacts OSDs configured with SSD as the backing device.

The number of threads per OSD shard must be factored in for calculating
the per shard bandwidth capacity of an OSD. This is because there could be
more than one worker thread configured per OSD shard and the OSD's
bandwidth capacity must be shared on a per shard thread basis.

For instance, currently, if the underlying OSD device is of SSD type, each
OSD shard is configured with two worker threads. Therefore, the per shard
thread bandwidth capacity must actually be calculated as:

  bandwidth_per_shard = OSD Bandwidth / (num_shards * threads_per_shard)

This results in the correct allocation of QoS reservation and limits to
the respective classes of service on the OSD. This has implications to
client QoS where reservation and limits could be something other than
MIN or MAX.

Without this change, and with SSD backing devices, each OSD shard thread
is incorrectly assigned double (2 worker threads) the intended reservation
and limit resulting in oversubscription of the OSD's bandwidth.

Changes:
--------
- Pass osd->get_num_op_threads() instead of num_shards to
  OpScheduler::make_scheduler()
- This is now passed as 'num_op_shard_threads' instead of 'num_shards' to
  the mClockScheduler constructor.
- In the mClockScheduler, osd_bandwidth_capacity_per_shard is renamed to
  osd_bandwidth_capacity_per_shard_thread and is calculated using the
  passed in num_op_shard_threads.
- osd_bandwidth_capacity_per_shard_thread is then used in
  update_from_config() to calculate the reservation and limit quantities.

Currently, for HDD based OSDs this change has no impact because the number
of threads per OSD shard is set to 1.

Fixes: https://tracker.ceph.com/issues/63197
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>