multi-mds/migrator: Fix assert during migration of referent inode
The dnl is expected to be null when the referent inode is migrated
during linkage. If the linkage is not null, the contents are expected
to be same. In this case, the remote was filled and we were trying
to re-link referent instead of just validating. Don't re-link, just
validate for referent inode.

Signed-off-by: Kotresh HR <khiremat@redhat.com>