qa/tasks/thrashosds-health: fix log-ignorelist of thrashosds-health.yaml
When any of secondary cannot respond from scrub reservation within a timeout period
due to various reasons (e.g., peering), it can be caused by PgScrubber timeout on primary
while reserving replicas.

There was a previous issues [1] that has been fixed already by pr [2].
However, it seems same problems occur in recent tests. [4-8]

In the previous commit [3], an log-ignorelist was updated for timeout logs.
After that, as the log message was changed from "timeout on replica" to "timeout on reserving replicas" [4],
the this message cannot filter out timeout warnings [5-9]. Because the log-ignorelist has’t been updated yet.

So to catch the "timeout on reserving replica" warning message,
the log-ignorelist needs to be fixed from "timeout on replica" to “timeout on reserving replicas”.

1. https://tracker.ceph.com/issues/59333
2. https://github.com/ceph/ceph/pull/52590/commits/ebccd68b97239d9923f08745ca6112bd2c2bfbee
3. https://github.com/ceph/ceph/pull/48545/commits/7730baa6c5b7c6ee579a1fb22c6336e6340614c7
4. https://github.com/ceph/ceph/commit/525478da3698cfe8a154c642dedd0570f1d52b37
5. https://pulpito.ceph.com/yuriw-2023-10-13_14:50:14-rados-wip-neorados-learning-from-experience-distro-default-smithi/7424512/
6. https://pulpito.ceph.com/ksirivad-2023-10-13_01:58:36-rados-wip-ksirivad-fix-63183-distro-default-smithi/7423913/
7. https://pulpito.ceph.com/yuriw-2023-10-05_14:23:09-rados-wip-yuri6-testing-2023-10-04-0901-distro-default-smithi/7411594
8. https://pulpito.ceph.com/myoungwon-2023-10-11_15:51:58-rados-wip-ceph-dedup-tool-snap-add-distro-default-smithi/7421887
9. https://pulpito.ceph.com/myoungwon-2023-10-11_15:51:58-rados-wip-ceph-dedup-tool-snap-add-distro-default-smithi/7421930/
* The tests [8, 9] are not including the previous bug fix commit [2],
  but it looks the same problem would have occurred even if the previous commits were included.

Signed-off-by: Sungmin Lee <sung_min.lee@samsung.com>
Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>