cephadm: add cephadm version --verbose option
The command `cephadm version --verbose` will emit JSON formatted output
detailing the metadata in the versions module as well as information
about what dependencies were bundled into the zipapp and a summary of
the contents of the zipapp. This should make it fairly easy to
determine the provenance of a given cephadm zipapp without needing to
unpack and analyze all the contents of the zip.

Signed-off-by: John Mulligan <jmulligan@redhat.com>