mgr/pg_autoscaler: add check for norecover flag
The PG autoscaler should not be active when the norecover flag is set.
As each newly split PG will need to be backfilled, it does not make sense to have the autoscaler active if the norecover flag is set.

Fixes: https://tracker.ceph.com/issues/63334
Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>
(cherry picked from commit 009832f07abe2bf051965097f550d8695e9cebd4)