src/osd/PrimaryLogPG.cc: add check for norecover flag in maybe_kick_recovery
Adds a check to see if norecover flag is set in maybe_kick_recovery().
maybe_kick_recovery() starts recovery for a particular object if there is an attempt to read a missing object or write to a degraded object.
However, this particular work-flow currently does not check the norecover flag before starting recovery for an object.

Fixes: bz#2134786
Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>