librgw: avoid illegal free after argv_to_vec
Remove the call to argv_to_vec(...) in librgw_create--because
the function probably is unsafe in general and reliably induced
an illegal free on destruction of what *should* have been a
temporary copy of one of its returned cstrings.

But also because the idea to call it was itself incorrect--it was
pushing the unsplit last argument back into args.

Fixes: https://tracker.ceph.com/issues/63380

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>