qa: test test_kill_mdstable for all mount types
It currently only runs for fuse and that hasn't had a successful instance in
months. The test skipped kernel mounts because you couldn't "kill" the mount
but that is no longer true with the addition of netns.

Fixes: https://tracker.ceph.com/issues/63432
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>
(cherry picked from commit 61577d9a9c6cfd4864c3961d856bc627dbfe0caf)