mgr/dashboard: update fedora link for dashboard-cephadm-e2e test
The old link points to a directory that now only contains a
README saying

---

ATTENTION
======================================
The contents of this directory have been moved to our archives available at:

http://archives.fedoraproject.org/pub/archive/fedora/

If you are having troubles finding something there please stop by #fedora-admin on irc.freenode.net

---

so I think we either need to start using a newer fedora or point
to the archive link. This commit is currently doing the latter.

Signed-off-by: Adam King <adking@redhat.com>