mgr/vol: don't always print success message for "volume rename" cmd
Currently 'FS volume "abc" was renamed to "xyz"' is printed even when
volume is absent on cluster, when only new volume name already exists,
when both volume names already exist and even when both volume names are
same. Instead in each of these cases a different message must be
printed.

Make this change and add tests for the same.

Signed-off-by: Rishabh Dave <ridave@redhat.com>