librbd: Append one journal event per object extent.
In the case where an image request is split across multiple object
extents and journaling is enabled, multiple journal events are appended.
Prior to this change, all object requests would wait for the last
journal event to complete, since journal events complete in order and
thus the last one completing implies that all prior journal events were
safe at that point.

While conceptually correct, journal events don't have a mechanism for
knowing how many requests might be associated with them. Thus, it's
entirely possible for the following sequence to occur:
1. An image request gets split into two object requests. Journal events
   are appended, one object request gets submitted and starts waiting on
   the last journal event (which also causes a C_CommitIOEvent to be
   instantiated against that journal event), but the other object
   request gets delayed due to an overlap.
2. Journaling completes, and the C_CommitIOEvent fires, which causes the
   event to be cleaned up.
3. The second object request from above is allowed to make progress; it
   tries to wait for the journal event that was just cleaned up which
   causes the assert in wait_event() to fire.

It might seem that the simple solution exists of allowing for
wait_event() to return early when the event no longer exists, and indeed
the code worked this way when originally written and was later changed
to enforce that the event existed. However, there appears to be another
problem here, which is that there is no C_CommitIOEvent instantiated for
the earlier journal event that were appended, which means they are
actually leaked and hang around forever. As far as I can tell, this
happens today whenever multiple journal events are appended.

The solution implemented here is to append a journal event for each
object extent that we're going to create a object request for. This
allows us to maintain a 1:1 relationship between object requests and
journal events, which in turn means that we can always safely look up a
journal event in the context of an object request and that a
C_CommitIOEvent is instantiated for each journal event.

Fixes: https://tracker.ceph.com/issues/63422
Signed-off-by: Joshua Baergen <jbaergen@digitalocean.com>