cephadm: add daemon_to_container utility function
Add the function daemon_to_container in the container_daemon_form.py
function. This function is roughly equivalent to `get_container` in
cephadm.py but entirely relies on subclasses implementing
ContainerDaemonForm methods and avoids special casing of daemon
types.

It duplicates the functionality provided by functions in cephadm.py with
regards to special podman specific behaviors. I feel that this is a
minor issue because due to recent refactoring those functions are now
the one- or two-liners that are copied here. I gated them behind
arguments `auto_podman_mounts` and `auto_podman_args` that default to
true in order to make the function more reusable - however I don't need
them that way and if it seems like extra complexity for no win, we can
remove them later.

Signed-off-by: John Mulligan <jmulligan@redhat.com>