osd: ECBackend::get_hash_info() takes external object size
This allows to reduce the interactions with `ObjectStore`
by letting callers `get_hash_info()` to provide the object
size from external source, from a cache in partiular.

The same idea will be reused in crimson but with an extra
benefit: no need for green threads within `submit_transaction()`
of `ECBackend`.

Signed-off-by: Radosław Zarzyński <rzarzyns@redhat.com>