doc/rados: format "initial troubleshooting"
Format the steps in the "Initial Troubleshooting" section of
doc/rados/troubleshooting/troubleshooting-mon.rst. A near-future PR (not
this one) will add context to this section and explain that the steps
described here are the first steps that you should undertake when you
determine that you have an unresponsive or down Monitor. This PR is
merely for formatting.

Signed-off-by: Zac Dover <zac.dover@proton.me>
(cherry picked from commit c581018caf626fa0dd50bd244766bfa9755c9a16)