add initial container file and github build action
This is only an initial implementation. It is not expected to work for
any branch except for Ceph 'main'. The build action is also disabled for
any pull request except ones that modify the container build files
themselves to make sure that any issues do not cause confusion with
ongoing or future feature PRs.

This PR focuses on creating a good starting point for a "Containerfile"
definition, tooling for building multi-architecture images using it, and
pushing the image to a temporary repository: quay.io/ceph/ceph-staging.

There are many TODO items noted in the files that should be followed up
on in future PRs.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>