Detect NFS availability without mgr crash
Ceph Dashboard needs to check availability of the NFS service
for the NFS tab.
However, the existing method caused the nfs mgr module to crash,
as ceph dashboard called `cluster_ls`,
which crashed if the orchestrator was not available.

Resolve the issue with two things:

1. Avoid crashing the module if orchestrator is not available.
The `cluster_ls` function now returns an empty list if no orchestrator.

2. Add a function to check if NFS is available.  This should be separate
from the `cluster_ls`, as we should avoid overloading the `cluster_ls`
function.  This way, we have a new function that can provide information
about whether the NFS + Orchestrator is available or not.

Closes-Bug: https://tracker.ceph.com/issues/56246

Closes-Bug: https://bugs.launchpad.net/charm-ceph-dashboard/+bug/2039955

Signed-off-by: Samuel Walladge <samuel.walladge@canonical.com>