src/osd/PeeringState.cc: update last_unstale properly
Problem:

When we update the `pg_stat` we don't
check whether the pg state is in `stale`.
Therefore, the attribute `last_unstale`
will always get updated even if the pg
state actually contains `stale`.

Solution:

Place a condition to only update
the attribute `last_unstale` when
we the pg truly doesn't have `stale`
in its state.

Fixes: https://tracker.ceph.com/issues/67777

Signed-off-by: Kamoltat <ksirivad@redhat.com>