mgr/vol: add helper methods in metadata_mgr.py
Add a helpers to check if config has a section and if section has option
set (or as it is said in codebase, key present section).

This will not only help in reducing typing effort and make code more
readable code but it will also bring more uniformity to the code.

Signed-off-by: Rishabh Dave <ridave@redhat.com>