qa: add centos_latest (9.stream) and ubuntu_20.04 yamls to supported-all-distro
A bug in Ceph MDS (MDS crash!) is seen with distos using a not-so-recent kernel
(5.4ish). This crash was first seen in quincy smoke run and the problematic
backport change was reverted. The smoke suite chooses a random distro for each
job, so to hit this bug, the appropriate distro needs to be (randomly) get chosen.

This change point the smoke suite to run against all supported distros.

This effects suites that point to supported-all-distro (powercycle) since it
bloats up the number of jobs. E.g., currently, without --subset, powercycle:osd

          INFO:teuthology.suite.run:0/336 jobs were filtered out.

vs

(with this change)

          Unable to schedule 560 jobs, too many jobs, when maximum 500 jobs allowed.

For smoke suite

          INFO:teuthology.suite.run:Scheduled 24 jobs in total.

vs

(with this change)

           INFO:teuthology.suite.run:Scheduled 120 jobs in total.

Eventually, with PR #46882, then testing kernel will no longer override the
distro kernel in fs suite, so we should get good coverage then.

Signed-off-by: Venky Shankar <vshankar@redhat.com>